pyproject.toml
[project]
name = "telegram-ai-provider"
version = "1.0.0"
description = "Telegram bot providing access to Claude API with vision and document support"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.15.0",
    "anthropic>=0.40.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
